Change Announcer API to return a Closable. (#652)
* Change Announcer API to return a Closable.

It's more convenient, for the purposes of linkerd's execution, for the
Announcer API to just return a Closable. The FutureAnnouncer helper bridges
Finagle's Future API to linkerd's.
